requirements.txt
tweepy~=4.10.0
requests~=2.28.0
python-dotenv~=0.20.0